Autodetect folder name in mod zip files
[The wiki](https://wiki.factorio.com/Tutorial:Mod_structure) says that
the a zip must contain a single folder and that that folder can be
named anything. As such, attempting to guess the folder name from
information we have (e.g. the mod name) will inevitably have failures
when mod authors don't follow the implicit practice of naming that
internal folder after the mod. The solution is to list out the
contents of the zipfile and feed those to `os.path.commonpath` to
obtain that single root directory.

Fixes #99